+++ sdds/classes.py
@@ -28,8 +28,8 @@ NUMTYPES_CAST = {"float": float, "double": float, "short": int,
                  "long": int, "llong": int, "char": str, "boolean": int}
 
 
-def get_dtype_str(type_: str, endianness: str = 'big'):
-    return f"{ENDIAN[endianness]}{NUMTYPES[type_]}"
+def get_dtype_str(type_: str, endianness: str = 'big', len: Any = ''):
+    return f"{ENDIAN[endianness]}{str(len)}{NUMTYPES[type_]}"
 
 
 ##############################################################################